address @tannewt changes: move and rename common files; remove PORT_HEAP_SIZE
