+++ ports/nrf/boards/pca10056/mpconfigboard.h
@@ -30,7 +30,6 @@
 #define MICROPY_HW_MCU_NAME         "nRF52840"
 #define MICROPY_PY_SYS_PLATFORM     "nRF52840-DK"
 
-#define PORT_HEAP_SIZE              (128 * 1024)
 #define CIRCUITPY_AUTORELOAD_DELAY_MS 500
 
 #define MICROPY_HW_LED_STATUS          (&pin_P0_13)
